feat: implement disk-based inverted index
